+++ approval-service/deploy/service.yaml
@@ -1,10 +1,10 @@
 apiVersion: apps/v1
 kind: Deployment
 metadata:
-  name: gatekeeper-service
+  name: approval-service
   namespace: keptn
   labels:
-    app.kubernetes.io/name: gatekeeper-service
+    app.kubernetes.io/name: approval-service
     app.kubernetes.io/instance: keptn
     app.kubernetes.io/part-of: keptn-keptn
     app.kubernetes.io/component: control-plane
@@ -12,13 +12,13 @@ metadata:
 spec:
   selector:
     matchLabels:
-      app.kubernetes.io/name: gatekeeper-service
+      app.kubernetes.io/name: approval-service
       app.kubernetes.io/instance: keptn
   replicas: 1
   template:
     metadata:
       labels:
-        app.kubernetes.io/name: gatekeeper-service
+        app.kubernetes.io/name: approval-service
         app.kubernetes.io/instance: keptn
         app.kubernetes.io/part-of: keptn-keptn
         app.kubernetes.io/component: control-plane
@@ -26,8 +26,8 @@ spec:
     spec:
       serviceAccountName: keptn-default
       containers:
-      - name: gatekeeper-service
-        image: keptn/gatekeeper-service:latest
+      - name: approval-service
+        image: keptn/approval-service:latest
         ports:
         - containerPort: 8080
         resources:
@@ -73,10 +73,10 @@ spec:
 apiVersion: v1
 kind: Service
 metadata:
-  name: gatekeeper-service
+  name: approval-service
   namespace: keptn
   labels:
-    app.kubernetes.io/name: gatekeeper-service
+    app.kubernetes.io/name: approval-service
     app.kubernetes.io/instance: keptn
     app.kubernetes.io/part-of: keptn-keptn
     app.kubernetes.io/component: control-plane
@@ -85,5 +85,5 @@ spec:
   - port: 8080
     protocol: TCP
   selector:
-    app.kubernetes.io/name: gatekeeper-service
+    app.kubernetes.io/name: approval-service
     app.kubernetes.io/instance: keptn
